webhook/requirements.txt
# Copyright 2023 Google LLC
#
# Licensed under the Apache License, Version 2.0 (the "License");
# you may not use this file except in compliance with the License.
# You may obtain a copy of the License at
#
#    https://www.apache.org/licenses/LICENSE-2.0
#
# Unless required by applicable law or agreed to in writing, software
# distributed under the License is distributed on an "AS IS" BASIS,
# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
# See the License for the specific language governing permissions and
# limitations under the License.

flask==2.3.2
# pyyaml installation is broken due to cython issues
# https://github.com/yaml/pyyaml/issues/724
pandas==1.3.5; python_version == '3.7'
pandas==2.0.1; python_version > '3.7'
pyyaml==5.3.1
functions-framework==3.4.0
google-auth==2.17.3
google-cloud-aiplatform[pipelines]>=1.25
google-cloud-bigquery==3.11.3
google-cloud-logging==3.5.0
google-cloud-storage==2.10.0
google-cloud-vision==3.4.4
requests==2.31.0
